fixed pattern search mesh handling
Fixed pattern search mesh handling. Previously changes to the mesh gridding
or current best/center point did not result in a translation/shifting of the
mesh origin to the new best/center point.  In the case of mesh expansion,
this could result in the mesh grid excluding the current best point that
polling occurs from (bad).  Also - in the case of continuous mesh usage for
the search step, the new best point could also be off the mesh.  Fixing this
required Adding a SetOrigin method to the mesh interface.

Other more minor changes include:

* Added pswarm.Vmax Option func

* Added test for non-zero origin on Infinite mesh.  This caught a bug where
  the final point wasn't being retranslated into original origin-space

* Renamed mesh.StepLimit to mesh.Integer and changed it to always just use an
  integral grid.

* optim.CacheEvaler now records the number of times its cache successfully
  prevented an objective evaluation.