fix skipping of rounded-to-center points during polling
Polling was intended to skip polling of points that, due to mesh-projection,
were identical to the starting/origin point of the poll.  This was coded wrong
and was comparing the unprojected point's distance to the starting point of
the poll.  It now correctly compares the distance from the projected point to
the poll origin point.